Dev mode for a better workflow

+++ primidi/source/primidi/menu.d
@@ -7,7 +7,7 @@ module primidi.menu;
 
 import std.stdio, std.conv, std.file, std.path, std.string;
 import atelier, grimoire, minuit;
-import primidi.loader, primidi.midi, primidi.gui, primidi.script;
+import primidi.config, primidi.loader, primidi.midi, primidi.gui, primidi.script;
 
 private {
 	enum _version = "0.2.0";
@@ -20,19 +20,19 @@ private {
 }
 
 bool processArguments(string[] args) {
-	if(args.length == 1)
+	if (args.length == 1)
 		return true;
 	string arg = args[1];
-	if(!arg.length)
+	if (!arg.length)
 		return false;
-	switch(arg) {
+	switch (arg) {
 	case "-v":
 	case "--version":
 		writeln("Primidi version " ~ _version);
 		break;
 	case "-h":
 	case "--help":
-		writeln("Primidi (c) 2016~2020 by Enalye.
+		writeln("Primidi (c) 2016~2021 by Enalye.
 Primidi is a free software and comes with ABSOLUTELY NO WARRANTY.
 
 Usage:
@@ -55,16 +55,18 @@ Where:
 }
 
 void setupApplication(string[] args) {
-	_lockFileName = buildNormalizedPath(dirName(thisExePath()), _lockFileName);
-	_msgFileName = buildNormalizedPath(dirName(thisExePath()), _msgFileName);
-	if(!exists(_lockFileName)) {
+	_lockFileName = buildNormalizedPath(getBasePath(), _lockFileName);
+	_msgFileName = buildNormalizedPath(getBasePath(), _msgFileName);
+	if (!exists(_lockFileName) || isDevMode()) {
 		// If the file exists, Primidi is already launched... or it crashed, we can never know.
 		_isMainApplication = true;
 	}
-	if(_isMainApplication) {
-		_lockFile = File(_lockFileName, "w");
-		_lockFile.lock();
-		
+	if (_isMainApplication) {
+		if (!isDevMode()) {
+			_lockFile = File(_lockFileName, "w");
+			_lockFile.lock();
+		}
+
 		enableAudio(false);
 		initializeMidiDevices();
 		createApplication(Vec2i(1280, 720 + 70), "Primidi");
@@ -72,17 +74,17 @@ void setupApplication(string[] args) {
 		setWindowMinSize(Vec2i(500, 200));
 		setWindowClearColor(Color.black);
 
-		const string iconPath = buildNormalizedPath(dirName(thisExePath()), "img", "icon.png");
-		if(exists(iconPath))
+		const string iconPath = buildNormalizedPath(getBasePath(), "img", "icon.png");
+		if (exists(iconPath))
 			setWindowIcon(iconPath);
 
-		if(!processArguments(args)) {
+		if (!processArguments(args)) {
 			closeLock();
 			return;
 		}
 		onStartupLoad(&onLoadComplete);
 
-		scope(exit) {
+		scope (exit) {
 			//We need to clean up the remaining threads.
 			closeLock();
 			stopMidiClock();
@@ -92,40 +94,43 @@ void setupApplication(string[] args) {
 		runApplication();
 		destroyApplication();
 	}
-	else if(processArguments(args)) {
-		if(_startingFilePath.length)
+	else if (processArguments(args)) {
+		if (_startingFilePath.length)
 			std.file.write(_msgFileName, _startingFilePath);
 	}
 }
 
 private void closeLock() {
-	if(!_isMainApplication)
+	if (!_isMainApplication)
 		return;
-	if(!exists(_lockFileName))
-		return;
-	_lockFile.close();
-	std.file.remove(_lockFileName);
+	if (!isDevMode()) {
+		if (!exists(_lockFileName))
+			return;
+		_lockFile.close();
+		std.file.remove(_lockFileName);
+	}
 }
 
 string receiveFilePath() {
-	if(!exists(_msgFileName) || isDir(_msgFileName))
+	if (!exists(_msgFileName) || isDir(_msgFileName))
 		return "";
 	string filePath;
 	try {
 		filePath = readText(_msgFileName);
 		std.file.remove(_msgFileName);
 	}
-	catch(Exception e) {}
-	if(!exists(filePath) || isDir(filePath))
+	catch (Exception e) {
+	}
+	if (!exists(filePath) || isDir(filePath))
 		return "";
 	const string ext = extension(filePath).toLower;
-	if(ext != ".mid" && ext != ".midi")
+	if (ext != ".mid" && ext != ".midi")
 		return "";
 	return filePath;
 }
 
 private void onLoadComplete() {
-    setDefaultFont(fetch!TrueTypeFont("Cascadia"));
+	setDefaultFont(fetch!TrueTypeFont("Cascadia"));
 	_mainGui = new MainGui(_startingFilePath);
 	onMainMenu();
 }
